+++ src/harmonist/formats/opus.py
@@ -37,6 +37,7 @@ read_track_title = _impl.read_track_title
 read_comment = _impl.read_comment
 read_duration_ms = _impl.read_duration_ms
 write_tags = _impl.write_tags
+write_cover = _impl.write_cover
 read_cover = _impl.read_cover